Refactor: Update tab navigation and home screen UI
This commit updates the tab navigation to use MaterialIcons and restyles the home screen with a new dashboard layout. It also cleans up unused code and improves overall UI consistency.
